requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
flask-limiter==3.8.0
bleach==6.1.0
pytest==7.4.3
pytest-flask==1.3.0